backend: Add LTE Base for EG25 and EC25
* Move EG25 backend to LTEBase
* Change EG25 and EC25 to use same backend since they are compatible in
  multiple aspects
* Risks of breaking QENG for EC25
