+++ backend/modem/__init__.py
@@ -3,5 +3,6 @@ from modem.at import ATCommander, ATCommand, ATDivider
 
 # Modem implementations
 from modem.adapters.quectel.lte_eg25_g import LTEEG25G
+from modem.adapters.quectel.lte_ec25 import LTEEC25
 
 __all__ = ["Modem", "ATCommander", "ATCommand", "ATDivider"]
